chore: Refactor placement of App.
