+++ app/src/main/java/no/usn/mob3000/MainActivity.kt
@@ -3,7 +3,6 @@ package no.usn.mob3000
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import no.usn.mob3000.ui.navigation.App
 import no.usn.mob3000.ui.theme.ChessbuddyTheme
 
 /**
